bug fixes and partial restoration
todo that I've noticed:
- Health bars and infestation bars missing on enemies and garages
- Menus are rolled back
- Upgrade menu is not controller compatible (I need a controller to test and fix this)

+++ Assets/Scripts/Player.cs
@@ -118,8 +118,7 @@ public class Player : MonoBehaviour
 
     public void UpdateHealthUI()
     {
-        healthFuelGauge.GetComponent<UIDialBehaviour>().UpdateGauge(health.health / health.maxHealth);
-        
+        healthFuelGauge.GetComponent<UIDialBehaviour>().UpdateGauge((float)health.health / (float)health.maxHealth);
         //if (healthBarFill != null)
         //{
         //    float t = (maxHealth > 0f) ? currentHealth / maxHealth : 0f;
